+++ crdesigner/ui/gui/controller/toolboxes/road_network_toolbox/intersections_controller.py
@@ -1,6 +1,5 @@
-from PyQt5.QtWidgets import QTableWidgetItem, QComboBox
+from PyQt5.QtWidgets import QTableWidgetItem
 from commonroad.scenario.intersection import IntersectionIncomingElement, Intersection
-from commonroad.scenario.traffic_sign import SupportedTrafficSignCountry
 
 from crdesigner.ui.gui.model.scenario_model import ScenarioModel
 from crdesigner.ui.gui.utilities.toolbox_ui import CheckableComboBox
@@ -9,7 +8,6 @@ from crdesigner.ui.gui.view.toolboxes.road_network_toolbox.road_network_toolbox_
     RoadNetworkToolboxUI
 from commonroad.scenario.traffic_sign import *
 from PyQt5.QtWidgets import *
-from PyQt5.QtCore import *
 
 
 class AddIntersectionController:
@@ -27,54 +25,14 @@ class AddIntersectionController:
         self.road_network_toolbox_ui.button_three_way_intersection.clicked.connect(
                 lambda: self.add_three_way_intersection())
         self.road_network_toolbox_ui.selected_intersection.currentTextChanged.connect(
-                lambda: self.update_intersection_information())
-        self.road_network_toolbox_ui.button_add_incoming.clicked.connect(lambda: self.add_incoming_to_table())
+                lambda: self.intersection_ui.update_intersection_information())
+        self.road_network_toolbox_ui.button_add_incoming.clicked.connect(lambda: self.intersection_ui.add_incoming_to_table())
         self.road_network_toolbox_ui.button_remove_incoming.clicked.connect(
             lambda: self.intersection_ui.remove_incoming())
         self.road_network_toolbox_ui.button_fit_intersection.clicked.connect(lambda: self.fit_intersection())
         self.road_network_toolbox_ui.button_add_intersection.clicked.connect(lambda: self.add_intersection())
         self.road_network_toolbox_ui.button_remove_intersection.clicked.connect(lambda: self.remove_intersection())
         self.road_network_toolbox_ui.button_update_intersection.clicked.connect(lambda: self.update_intersection())
-
-    def update_intersection_information(self):
-        """
-        Updates information of intersection widget based on intersection ID selected by the user.
-        """
-        if self.road_network_controller.mwindow.play_activated:
-            self.road_network_controller.text_browser.append("Please stop the animation first.")
-            return
-
-        if self.road_network_toolbox_ui.selected_intersection.currentText() not in ["", "None"]:
-            selected_intersection_id = int(self.road_network_toolbox_ui.selected_intersection.currentText())
-            intersection = self.scenario_model.find_intersection_by_id(selected_intersection_id)
-            self.road_network_toolbox_ui.intersection_incomings_table.setRowCount(0)
-            incoming_ids = [str(inc.incoming_id) for inc in intersection.incomings]
-            for incoming in intersection.incomings:
-                self.add_incoming_to_table(False, incoming_ids)
-                num_rows = self.road_network_toolbox_ui.intersection_incomings_table.rowCount()
-                self.road_network_toolbox_ui.intersection_incomings_table.setItem(num_rows - 1, 0, QTableWidgetItem(
-                    str(incoming.incoming_id)))
-                self.road_network_toolbox_ui.intersection_incomings_table.cellWidget(num_rows - 1, 1).set_checked_items(
-                        [str(la_id) for la_id in incoming.incoming_lanelets])
-                self.road_network_toolbox_ui.intersection_incomings_table.cellWidget(num_rows - 1, 2).set_checked_items(
-                        [str(la_id) for la_id in incoming.successors_left])
-                self.road_network_toolbox_ui.intersection_incomings_table.cellWidget(num_rows - 1, 3).set_checked_items(
-                        [str(la_id) for la_id in incoming.successors_straight])
-                self.road_network_toolbox_ui.intersection_incomings_table.cellWidget(num_rows - 1, 4).set_checked_items(
-                        [str(la_id) for la_id in incoming.successors_right])
-                index = self.road_network_toolbox_ui.intersection_incomings_table.cellWidget(num_rows - 1, 5).findText(
-                        str(incoming.left_of))
-                self.road_network_toolbox_ui.intersection_incomings_table.cellWidget(num_rows - 1, 5).setCurrentIndex(
-                        index)
-            self.road_network_toolbox_ui.intersection_crossings.set_checked_items(
-                    [str(cr) for cr in intersection.crossings])
-
-            self.road_network_toolbox_ui.intersection_lanelet_to_fit.clear()
-            current_text = self.road_network_toolbox_ui.selected_intersection.currentText()
-            self.road_network_toolbox_ui.intersection_lanelet_to_fit \
-                .addItems(["None"] + [str(item) for item in
-                                      self.scenario_model.collect_incoming_lanelet_ids_from_intersection(current_text)])
-            self.road_network_toolbox_ui.intersection_lanelet_to_fit.setCurrentIndex(0)
 
     def add_four_way_intersection(self):
         """
@@ -223,46 +181,3 @@ class AddIntersectionController:
             selected_intersection_id = int(self.road_network_toolbox_ui.selected_intersection.currentText())
             self.scenario_model.update_intersection(selected_intersection_id)
             self.add_intersection(selected_intersection_id)
-            """
-            old code if there needs rework
-            self.set_default_road_network_list_information()
-            self.callback(self.current_scenario)
-            """
-
-    def add_incoming_to_table(self, new_incoming: bool = True, incoming_ids: List[str] = None):
-        """
-        Adds a row to the intersection incoming table.
-        Only a default entry is created the user has to specify the incoming afterward manually.
-
-        @param new_incoming: Boolean indicating whether this will be an new incoming or from a new intersection
-        @param incoming_ids: List of available incoming IDs.
-        """
-        if self.road_network_controller.mwindow.play_activated:
-            self.road_network_controller.text_browser.append("Please stop the animation first.")
-            return
-
-        if not self.scenario_model.scenario_created():
-            self.road_network_controller.text_browser.append("_Warning:_ Create a new file")
-            return
-
-        num_rows = self.road_network_toolbox_ui.intersection_incomings_table.rowCount()
-        self.road_network_toolbox_ui.intersection_incomings_table.insertRow(num_rows)
-        lanelet_ids = [str(la_id) for la_id in self.scenario_model.collect_lanelet_ids()]
-        if new_incoming:
-            self.road_network_toolbox_ui.intersection_incomings_table.setItem(
-                    num_rows,
-                    0,QTableWidgetItem(str(self.scenario_model.generate_object_id())))
-        combo_box_lanelets = CheckableComboBox()
-        combo_box_lanelets.addItems(lanelet_ids)
-        self.road_network_toolbox_ui.intersection_incomings_table.setCellWidget(num_rows, 1, combo_box_lanelets)
-        combo_box_successors_left = CheckableComboBox()
-        combo_box_successors_left.addItems(lanelet_ids)
-        self.road_network_toolbox_ui.intersection_incomings_table.setCellWidget(num_rows, 2, combo_box_successors_left)
-        combo_box_successors_straight = CheckableComboBox()
-        combo_box_successors_straight.addItems(lanelet_ids)
-        self.road_network_toolbox_ui.intersection_incomings_table.setCellWidget(num_rows, 3,
-                                                                                combo_box_successors_straight)
-        combo_box_successors_right = CheckableComboBox()
-        combo_box_successors_right.addItems(lanelet_ids)
-        self.road_network_toolbox_ui.intersection_incomings_table.setCellWidget(num_rows, 4, combo_box_successors_right)
-        self.intersection_ui.update_left_of_combobox(incoming_ids)